build relationship between user and tasks

+++ src/tasks/tasks.service.ts
@@ -4,6 +4,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -12,39 +13,40 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getAllTasks(): Promise<Task[]> {
-    const tasks = await this.taskRepository
-    .createQueryBuilder("task")
-    .getMany();
-    return tasks;
+  async getAllTasks(user: User): Promise<Task[]> {
+    return await this.taskRepository
+      .createQueryBuilder("task").andWhere('task.userId = :userId', { userId: user.id })
+      .getMany();
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(user: User, id: number): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: {id, userId: user.id} });
     if (!found) {
       throw new NotFoundException(`不存在ID为${id}的Task`);
     }
     return found;
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(user: User, id: number, status: TaskStatus): Promise<Task> {
+    const task = await this.getTaskById(user, id);
     task.status = status;
     await task.save();
     return task;
   }
 
-  async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskById(user: User, id: number): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if(!result.affected) {
       throw new NotFoundException(`不存在ID为${id}的Task`);
     }
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
+  async createTask(user: User, createTaskDto: CreateTaskDto): Promise<Task> {
     const {title, desc} = createTaskDto;
     const task = this.taskRepository.create({title, desc, status: TaskStatus.open})
+    task.user = user;
     await task.save();
+    delete task.user
     return task;
   }
 }
